Change factorial return type to unsigned long long
Updated the factorial function to return unsigned long long instead of unsigned int to support larger results and prevent overflow for bigger input values.

+++ math/factorial.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-unsigned int factorial(unsigned int n)
+unsigned long long factorial(unsigned int n)
 {
 
     return n == 0 ? 1 : n * factorial(n - 1);
